getting confirmation and email to work
Working in hello1.rb - have been able to make the email pull in all the
details, check it. I can't work out how to do an equation to get it to
add the cost of venue and chef together but will keep trying. Also it's
stopped posting to twitter - how annoying.